Made the launchpad and added some Bootsrap classnames
