+++ src/scss/app.scss
@@ -1,0 +1,4 @@
+.wrapper{
+    height: 100vh;
+    width: 100%;
+}
