+++ src/scss/index.scss
@@ -1,0 +1,4 @@
+body{
+    margin: 0;
+    padding: 0;
+}
